feat: reservationCost를 integer에서 long으로 변경

+++ src/main/java/com/dentallink/domain/hospital/dto/request/HospitalCreateRequest.java
@@ -8,7 +8,7 @@ public record HospitalCreateRequest (
     String hospitalAddress,
     Boolean hospitalIsOpen,
     String doctorName,
-    Integer reservationCost,
+    Long reservationCost,
     LocalTime openTime,
     LocalTime closeTime,
     LocalTime breakStart,
